refactor: split changeset generation from file creation
