+++ packages/changelog/src/generateChangeset.test.ts
@@ -8,9 +8,9 @@ import {
 } from '@monodeploy/test-utils'
 import { PortablePath } from '@yarnpkg/fslib'
 
-import { writeChangesetFile } from '.'
+import { generateChangeset } from '.'
 
-describe('writeChangesetFile', () => {
+describe('generateChangeset', () => {
     let workspacePath: PortablePath
 
     beforeEach(async () => {
@@ -37,7 +37,7 @@ describe('writeChangesetFile', () => {
         }
         const context = await setupContext(cwd)
 
-        const changeset = await writeChangesetFile({
+        const changeset = await generateChangeset({
             config,
             context,
             previousTags: new Map([
