+++ packages/changelog/src/index.ts
@@ -1,5 +1,6 @@
+import generateChangeset from './generateChangeset'
 import prependChangelogFile from './prependChangelogFile'
 import resolveConventionalConfig from './resolveConventionalConfig'
 import writeChangesetFile from './writeChangesetFile'
 
-export { prependChangelogFile, writeChangesetFile, resolveConventionalConfig }
+export { prependChangelogFile, writeChangesetFile, resolveConventionalConfig, generateChangeset }
